requirements.txt
# requirements file for Python libraries
# install with 'pip install -r requirements.txt'
# https://note.nkmk.me/en/python-pip-install-requirements/

SPARQLWrapper
python-benedict
ipypublish
pandas
sparql_dataframe
matplotlib
plotly
pillow
thothlibrary==0.16.0
VizKG
markdownify
regex